+++ packages/sdk/src/proveOperation.ts
@@ -92,9 +92,9 @@ async function proveJoinSplit(
         nullifierB: baseJoinSplit.nullifierB,
         senderCommitment,
         opDigest,
-        encododedAssetAddrWithSignBits:
-          proofInputs.encodedAssetAddrWithSignBits,
-        encodedAssetID: proofInputs.encodedAssetId,
+        pubEncododedAssetAddrWithSignBits:
+          proofInputs.pubEncodedAssetAddrWithSignBits,
+        pubEncodedAssetID: proofInputs.pubEncodedAssetId,
         refundAddrH1CompressedY: refundAddrH1CompressedY,
         refundAddrH2CompressedY: refundAddrH2CompressedY,
       },